nimble/ll: Fix extended advertising reports handling
We may run out of buffers for HCI advertising reports if scanning a lot
of devices at the same time and/or host does not handle these events
fast enough to return buffers to pool. This is handled properly if we
need to fragment advertising data and fail to allocate buffer for 2nd
or subsequent events (we'll mark data as trimmed in current event), but
in case we cannot allocate buffer for 1st event it will just be dropped
without marking data as trimmed. As a result, another PDU scanned from
the same advertiser may be handled by host as continuation of previous
data which results in corrupted reassembled data and there is no way
for host to figure this out.

This patch changes extended advertising reports handling to make sure
data are always marked as trimmed when we run out of event buffers. We
now always keep one pre-allocated event in aux_data so we can always
send at least one event that will mark data as trimmed. If we cannot
allocate 1st event for chain, it will be just dropped which is ok since
host will not event know about this chain.